Remove PROPOSE_VIRTUAL
