+++ packages/cf-funding-protocol-contracts/package.json
@@ -33,7 +33,7 @@
   ],
   "devDependencies": {
     "@counterfactual/cf-adjudicator-contracts": "0.0.6",
-    "@counterfactual/types": "0.0.39",
+    "@counterfactual/types": "0.0.42",
     "@counterfactual/typescript-typings": "0.1.2",
     "@types/node": "12.7.4",
     "@types/chai": "4.2.1",
