+++ packages/node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@counterfactual/node",
-  "version": "0.2.79",
+  "version": "0.2.80",
   "main": "dist/index.js",
   "iife": "dist/index.iife.js",
   "types": "dist/src/index.d.ts",
@@ -58,7 +58,7 @@
     "@counterfactual/cf-funding-protocol-contracts": "0.0.10",
     "@counterfactual/cf.js": "0.2.6",
     "@counterfactual/firebase-client": "0.0.6",
-    "@counterfactual/types": "0.0.39",
+    "@counterfactual/types": "0.0.42",
     "ethers": "4.0.36",
     "eventemitter3": "^4.0.0",
     "loglevel": "^1.6.1",
