+++ packages/node/src/api.ts
@@ -1,6 +1,5 @@
 import {
   handleReceivedProposalMessage,
-  handleReceivedProposeVirtualMessage,
   handleRejectProposalMessage
 } from "./message-handling/handle-node-message";
 import { handleReceivedProtocolMessage } from "./message-handling/handle-protocol-message";
@@ -20,7 +19,6 @@ import {
   InstallAppInstanceController,
   InstallVirtualAppInstanceController,
   ProposeInstallAppInstanceController,
-  ProposeInstallVirtualAppInstanceController,
   RejectInstallController,
   TakeActionController,
   UninstallController,
@@ -42,7 +40,6 @@ const controllers = [
   InstallAppInstanceController,
   InstallVirtualAppInstanceController,
   ProposeInstallAppInstanceController,
-  ProposeInstallVirtualAppInstanceController,
   RejectInstallController,
   TakeActionController,
   UninstallController,
@@ -99,7 +96,12 @@ export const createRpcRouter = (requestHandler: RequestHandler) =>
 
 export const eventNameToImplementation = {
   [NODE_EVENTS.PROPOSE_INSTALL]: handleReceivedProposalMessage,
-  [NODE_EVENTS.PROPOSE_INSTALL_VIRTUAL]: handleReceivedProposeVirtualMessage,
+  [NODE_EVENTS.PROPOSE_INSTALL_VIRTUAL]: (rh, p) => {
+    console.warn(
+      "WARNING: PROPOSE_INSTALL_VIRTUAL is deprecated. Replace usage with PROPOSE_INSTALL."
+    );
+    return handleReceivedProposalMessage(rh, p);
+  },
   [NODE_EVENTS.PROTOCOL_MESSAGE_EVENT]: handleReceivedProtocolMessage,
   [NODE_EVENTS.REJECT_INSTALL]: handleRejectProposalMessage,
   [NODE_EVENTS.REJECT_INSTALL_VIRTUAL]: handleRejectProposalMessage
