+++ packages/node/src/message-handling/handle-node-message.ts
@@ -1,15 +1,6 @@
-import {
-  getNextNodeAddress,
-  getOrCreateStateChannelBetweenVirtualAppParticipants
-} from "../methods/app-instance/propose-install-virtual/operation";
 import { AppInstanceProposal } from "../models";
 import { RequestHandler } from "../request-handler";
-import {
-  NODE_EVENTS,
-  ProposeMessage,
-  ProposeVirtualMessage,
-  RejectProposalMessage
-} from "../types";
+import { ProposeMessage, RejectProposalMessage } from "../types";
 import { getCreate2MultisigAddress } from "../utils";
 
 export async function handleReceivedProposalMessage(
@@ -29,7 +20,11 @@ export async function handleReceivedProposalMessage(
     networkContext.MinimumViableMultisig
   );
 
-  const stateChannel = await store.getStateChannel(multisigAddress);
+  const stateChannel = await store.getOrCreateStateChannelBetweenVirtualAppParticipants(
+    multisigAddress,
+    publicIdentifier,
+    proposedByIdentifier
+  );
 
   await store.addAppInstanceProposal(
     stateChannel,
@@ -58,55 +53,4 @@ export async function handleRejectProposalMessage(
     data: { appInstanceId }
   } = receivedRejectProposalMessage;
   await store.removeAppInstanceProposal(appInstanceId);
-}
-
-export async function handleReceivedProposeVirtualMessage(
-  requestHandler: RequestHandler,
-  receivedProposeMessage: ProposeVirtualMessage
-) {
-  const { store, networkContext } = requestHandler;
-
-  const {
-    data: { params, proposedByIdentifier }
-  } = receivedProposeMessage;
-
-  const {
-    intermediaryIdentifier,
-    proposedToIdentifier,
-    responderDeposit,
-    responderDepositTokenAddress,
-    initiatorDeposit,
-    initiatorDepositTokenAddress
-  } = params;
-
-  const multisigAddress = getCreate2MultisigAddress(
-    [proposedByIdentifier, proposedToIdentifier],
-    networkContext.ProxyFactory,
-    networkContext.MinimumViableMultisig
-  );
-
-  const stateChannel = await getOrCreateStateChannelBetweenVirtualAppParticipants(
-    multisigAddress,
-    proposedByIdentifier,
-    proposedToIdentifier,
-    intermediaryIdentifier,
-    store,
-    networkContext
-  );
-
-  await store.addVirtualAppInstanceProposal(
-    new AppInstanceProposal(
-      {
-        ...params,
-        proposedByIdentifier,
-        initiatorDeposit: responderDeposit,
-        initiatorDepositTokenAddress: responderDepositTokenAddress!,
-        responderDeposit: initiatorDeposit,
-        responderDepositTokenAddress: initiatorDepositTokenAddress!
-      },
-      stateChannel
-    )
-  );
-
-  await store.saveStateChannel(stateChannel.bumpProposedApps());
 }
