+++ packages/node/src/methods/app-instance/propose-install/controller.ts
@@ -24,6 +24,7 @@ import { createProposedAppInstance } from "./operation";
  */
 export default class ProposeInstallController extends NodeController {
   @jsonRpcMethod(Node.RpcMethodName.PROPOSE_INSTALL)
+  @jsonRpcMethod(Node.RpcMethodName.PROPOSE_INSTALL_VIRTUAL)
   public executeMethod = super.executeMethod;
 
   protected async getRequiredLockNames(
@@ -75,7 +76,11 @@ export default class ProposeInstallController extends NodeController {
     const responderDepositTokenAddress =
       responderDepositTokenAddressParam || CONVENTION_FOR_ETH_TOKEN_ADDRESS;
 
-    const stateChannel = await store.getStateChannel(multisigAddress);
+    const stateChannel = await store.getOrCreateStateChannelBetweenVirtualAppParticipants(
+      multisigAddress,
+      myIdentifier,
+      proposedToIdentifier
+    );
 
     assertSufficientFundsWithinFreeBalance(
       stateChannel,
@@ -135,6 +140,8 @@ function assertSufficientFundsWithinFreeBalance(
   tokenAddress: string,
   depositAmount: BigNumber
 ) {
+  if (!channel.hasFreeBalance) return;
+
   const freeBalanceForToken = channel
     .getFreeBalanceClass()
     .getBalance(tokenAddress, xkeyKthAddress(publicIdentifier, 0));
