+++ packages/node/test/integration/install-virtual.spec.ts
@@ -2,7 +2,7 @@ import { NetworkContextForTestSuite } from "@counterfactual/local-ganache-server
 import { Node as NodeTypes } from "@counterfactual/types";
 
 import { Node } from "../../src";
-import { NODE_EVENTS, ProposeVirtualMessage } from "../../src/types";
+import { NODE_EVENTS, ProposeMessage } from "../../src/types";
 
 import { setup, SetupContext } from "./setup";
 import {
@@ -51,9 +51,9 @@ describe("Node method follows spec - proposeInstallVirtual", () => {
 
         nodeC.once(
           NODE_EVENTS.PROPOSE_INSTALL_VIRTUAL,
-          async (msg: ProposeVirtualMessage) => {
+          async (msg: ProposeMessage) => {
             const { appInstanceId } = msg.data;
-            const { intermediaryIdentifier } = msg.data.params;
+
             const [proposedAppNodeA] = await getProposedAppInstances(nodeA);
             const [proposedAppNodeC] = await getProposedAppInstances(nodeC);
 
@@ -70,7 +70,7 @@ describe("Node method follows spec - proposeInstallVirtual", () => {
             expect(proposedAppNodeA.identityHash).toEqual(
               proposedAppNodeC.identityHash
             );
-            installTTTVirtual(nodeC, appInstanceId, intermediaryIdentifier);
+            installTTTVirtual(nodeC, appInstanceId, nodeB.publicIdentifier);
           }
         );
