+++ packages/node/test/integration/reject-install-virtual.spec.ts
@@ -2,7 +2,7 @@ import { NetworkContextForTestSuite } from "@counterfactual/local-ganache-server
 import { Node as NodeTypes } from "@counterfactual/types";
 
 import { Node } from "../../src";
-import { NODE_EVENTS, ProposeVirtualMessage } from "../../src/types";
+import { NODE_EVENTS, ProposeMessage } from "../../src/types";
 
 import { setup, SetupContext } from "./setup";
 import {
@@ -41,7 +41,7 @@ describe("Node method follows spec - rejectInstallVirtual", () => {
 
         nodeC.on(
           NODE_EVENTS.PROPOSE_INSTALL_VIRTUAL,
-          async (msg: ProposeVirtualMessage) => {
+          async (msg: ProposeMessage) => {
             const { appInstanceId } = msg.data;
             const [proposedAppInstanceA] = await getProposedAppInstances(nodeA);
             const [proposedAppInstanceC] = await getProposedAppInstances(nodeC);
