+++ packages/types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@counterfactual/types",
-  "version": "0.0.41",
+  "version": "0.0.42",
   "description": "TypeScript typings for common Counterfactual types",
   "main": "dist/index.js",
   "module": "dist/index.esm.js",
